fix(maps): Make skybox rotation speed Double

+++ src/main/kotlin/jp/assasans/protanki/server/ResourceConverter.kt
@@ -100,7 +100,7 @@ data class ServerMapVisual(
   @Json val nearLimit: Int,
 
   @Json val gravity: Int,
-  @Json val skyboxRevolutionSpeed: Int,
+  @Json val skyboxRevolutionSpeed: Double,
   @Json val ssaoColor: Int,
 
   @Json val dustAlpha: Double,
